+++ packages/auth/__tests__/providers/cognito/signInWithRedirect.test.ts
@@ -47,11 +47,12 @@ jest.mock('@aws-amplify/core', () => {
 		ConsoleLogger: jest.fn(),
 	};
 });
+
 jest.mock('../../../src/providers/cognito/utils/signInHelpers');
+
 jest.mock('../../../src/providers/cognito/utils/oauth', () => ({
 	...jest.requireActual('../../../src/providers/cognito/utils/oauth'),
 	completeOAuthFlow: jest.fn(),
-	handleFailure: jest.fn(),
 	generateCodeVerifier: jest.fn(),
 	generateState: jest.fn(),
 }));
@@ -70,6 +71,7 @@ jest.mock('../../../src/providers/cognito/utils/oauth/oAuthStore', () => ({
 		clearOAuthInflightData: jest.fn(),
 	} as OAuthStore,
 }));
+jest.mock('../../../src/providers/cognito/utils/oauth/handleFailure');
 jest.mock('../../../src/providers/cognito/utils/oauth/createOAuthError');
 jest.mock('../../../src/utils');
 
@@ -188,6 +190,26 @@ describe('signInWithRedirect', () => {
 				);
 			});
 		});
+
+		it('invokes handleFailure when user cancels the oauth flow', async () => {
+			const error = new Error('OAuth flow was cancelled.')
+			const mockOpenAuthSessionResult = {
+				type: undefined,
+			};
+			mockCreateOAuthError.mockReturnValueOnce(error);
+			mockOpenAuthSession.mockResolvedValueOnce(mockOpenAuthSessionResult);
+			oAuthStore.loadOAuthInFlight = jest.fn().mockResolvedValueOnce(true);
+			const currentAddEventlistener = window.addEventListener;
+			window.addEventListener = jest.fn((event: string, cb: any) => {
+				cb({ persisted: true });
+			});
+
+			await signInWithRedirect({ provider: 'Google' });
+			expect(mockCreateOAuthError).toHaveBeenCalledTimes(1);
+			expect(mockHandleFailure).toHaveBeenCalledWith(error);
+		    
+			window.addEventListener = currentAddEventlistener;
+		})
 	});
 
 	describe('specifications on react-native', () => {
@@ -275,6 +297,7 @@ describe('signInWithRedirect', () => {
 
 			expect(oAuthStore.storeOAuthInFlight).toHaveBeenCalledTimes(0);
 		});
+
 	});
 
 	describe('errors', () => {
@@ -304,5 +327,6 @@ describe('signInWithRedirect', () => {
 
 			await expect(signInWithRedirect()).rejects.toThrow(mockError);
 		});
+
 	});
 });